Added Card ScriptableObject and the UI display
We added CardDisplay.cs and Card.cs where we can have an Object in our scene and can display cards as UI elements like we want. Works fine for now no real logic behind it for now. But we can basically create as much cards as we want but with no Functionality