Currently Implementing Deck and Hands
We can not see our Hand in this state BUT we do have a deck that is set up to have 30 cards. These cards are defined inside the Deck scriptable Object. We also have some functions in our DeckManager where we draw cards and stuff and also got a hand that theoretically is drawing hands but we dont slot them right now and cant see them ingame