3D Cards functioning now - Simple CardDrawPhase
Both players have a hand and a deck and for now only 1 player can draw a card but still we can focking do it. This is a good state. The phases are a bit fucked up as to what phase is actually is and how the Index gets updated. Have to improve that. It fucntions but it is awkward