Made Card Handling smoother and nicer
There is still a bit of an "bug" where the card is returning to the hand for a millisecond because our UpdateLayout is calling the card and the draggable is calling it and depending on who is updating faster it gets placed there. Its not that big of a deal but should be fixed somewhen.